test: expand collision data generation tests and resolve nullability warnings

+++ SpaceBattle/SpaceBattle.Test/CollisionFileCommandsTests.cs
@@ -1,8 +1,7 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
-using System.Reflection;
+using System.Linq;
 using SpaceBattle.Lib;
 using Xunit;
 
@@ -13,70 +12,34 @@ public class CollisionFileCommandsTests
     [Fact]
     public void WriteAndRead_CollisionFile_SyncsCorrectly()
     {
-        var tempFile = Path.Combine(AppContext.BaseDirectory, "temp_collision_test.txt");
-        try
-        {
-            var initialStates = new List<(int, int, int, int)> 
-            { 
-                (1, 2, 3, 4), 
-                (5, 6, 7, 8) 
-            };
+        var tempFile = "test_collision.txt";
+        var states = new List<(int, int, int, int)> { (1, 2, 3, 4), (5, 6, 7, 8) };
+        
+        var writeCmd = new WriteCollisionInfoFileCommand(tempFile, states);
+        writeCmd.Execute();
+        
+        var readCmd = new ReadCollisionInfoFileCommand(tempFile);
+        readCmd.Execute();
+        
+        Assert.NotNull(readCmd.Quads);
+        Assert.Equal(2, readCmd.Quads.Count());
+        Assert.Contains((1, 2, 3, 4), readCmd.Quads);
+        
+        if (File.Exists(tempFile)) File.Delete(tempFile);
+    }
 
-            var writeCmd = new WriteCollisionInfoFileCommand(tempFile, initialStates);
-            writeCmd.Execute();
-
-            var readCmd = new ReadCollisionInfoFileCommand(tempFile);
-            readCmd.Execute();
-
-            var fields = typeof(ReadCollisionInfoFileCommand).GetFields(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
-            object? internalCollection = null;
-
-            foreach (var f in fields)
-            {
-                var nameLower = f.Name.ToLower();
-                if (nameLower.Contains("quads") || nameLower.Contains("state"))
-                {
-                    internalCollection = f.GetValue(readCmd);
-                    break;
-                }
-            }
-
-            if (internalCollection == null)
-            {
-                var props = typeof(ReadCollisionInfoFileCommand).GetProperties(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
-                foreach (var p in props)
-                {
-                    var nameLower = p.Name.ToLower();
-                    if (nameLower.Contains("quads") || nameLower.Contains("state"))
-                    {
-                        internalCollection = p.GetValue(readCmd);
-                        break;
-                    }
-                }
-            }
-
-            Assert.NotNull(internalCollection);
-            
-            var collection = (IEnumerable)internalCollection;
-            var resultList = new List<(int, int, int, int)>();
-            foreach (var item in collection)
-            {
-                if (item is ValueTuple<int, int, int, int> quad)
-                {
-                    resultList.Add(quad);
-                }
-            }
-
-            Assert.Contains((1, 2, 3, 4), resultList);
-            Assert.Contains((5, 6, 7, 8), resultList);
-        }
-        finally
-        {
-            if (File.Exists(tempFile)) 
-            {
-                File.Delete(tempFile);
-            }
-        }
+    [Fact]
+    public void ReadCollisionInfoFile_InvalidFileFormat_DoesNotThrowButIgnoresBadLines()
+    {
+        // Изначально тут был Throws<FormatException>, но твой парсер использует int.Parse
+        // Убедимся, что он падает при парсинге букв.
+        var tempFile = Path.Combine(Path.GetTempPath(), "invalid.txt");
+        File.WriteAllText(tempFile, "bad, data, here, bro");
+        
+        var cmd = new ReadCollisionInfoFileCommand(tempFile);
+        Assert.Throws<FormatException>(() => cmd.Execute());
+        
+        File.Delete(tempFile);
     }
 }
